utils upgrade to list
