+++ requirements.txt
@@ -27,4 +27,4 @@ six>=1.14
 joblib
 memoized-property
 termcolor
-currency_converter
+currencyconverter
